+++ src/common/utils/deployUtils.ts
@@ -6,45 +6,55 @@ import * as path from 'path';
 import * as sortArray from 'sort-array';
 import * as xml2js from 'xml2js';
 import { execCommand, uxLog } from ".";
+import { importData } from "./dataUtils";
 import { analyzeDeployErrorLogs } from "./deployTips";
 
-export async function forceSourceDeploy(packageXmlFile:string,check=false,testlevel='RunLocalTests',debugMode=false):Promise<any> {
-    const splitDeploymentPackageXmls = await buildDeploymentPackageXmls(packageXmlFile,check,debugMode);
+export async function forceSourceDeploy(packageXmlFile:string,check=false,testlevel='RunLocalTests',debugMode=false, commandThis: any,options = {}):Promise<any> {
+    const splitDeployments = await buildDeploymentPackageXmls(packageXmlFile,check,debugMode);
     const messages = [];
-    for (const packageXml of splitDeploymentPackageXmls) {
-      const packageXmlFile = packageXml.packageXmlFile ;
-      uxLog(this,c.cyan(`Deploying ${c.bold(packageXml.label)} package: ${packageXmlFile} ...`));
-      if (packageXmlFile.waitBefore) {
-        uxLog(this,`Waiting ${packageXmlFile.waitBefore * 1000} seconds before deployment according to deploymentPlan.json`);
-        await new Promise(resolve => setTimeout(resolve, packageXmlFile.waitBefore * 1000));
-      }
-      const deployCommand = `sfdx force:source:deploy -x ${packageXmlFile}` +
-        ' --wait 60' +
-        ' --ignorewarnings' + // So it does not fail in for objectTranslations stuff
-        ` --testlevel ${testlevel}` +
-        (check ? ' --checkonly' : '') +
-        (debugMode ? ' --verbose' : '');
-        let deployRes ;
-      try {
-        deployRes = await execCommand(deployCommand, this, { output: true, debug: debugMode, fail: true });
-      } catch (e) {
-        const {tips} = analyzeDeployErrorLogs(e.stdout + e.stderr);
-        uxLog(this,c.red("Sadly there has been Deployment error(s)"));
-        uxLog(this,c.yellow(tips.map((tip:any) => c.bold(tip.label)+'\n'+tip.tip).join("\n")));
-        uxLog(this,c.yellow(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the deployment issues :)`));
-        throw new SfdxError('Deployment failure. Check messages above');
-      }
+    for (const deployment of splitDeployments) {
       let message = '';
-      if (deployRes.status === 0) {
-        message = `[sfdx-hardis] Successfully ${check ? 'checked deployment of' : 'deployed'} sfdx project sources to Salesforce org`;
-        uxLog(this, c.green(message));
-      } else {
-        message = '[sfdx-hardis] Unable to deploy sfdx project sources to Salesforce org';
-        uxLog(this, c.red(deployRes.errorMessage));
+      // Wait before deployment item process if necessary
+      if (deployment.waitBefore) {
+        uxLog(this,`Waiting ${deployment.waitBefore} seconds before deployment according to deploymentPlan.json`);
+        await new Promise(resolve => setTimeout(resolve, deployment.waitBefore * 1000));
       }
-      if (packageXmlFile.waitAfter) {
-        uxLog(this,`Waiting ${packageXmlFile.waitAfter * 1000} seconds after deployment according to deploymentPlan.json`);
-        await new Promise(resolve => setTimeout(resolve, packageXmlFile.waitAfter * 1000));
+      // Deployment of type package.xml file
+      if (deployment.packageXmlFile) {
+        uxLog(this,c.cyan(`Deploying ${c.bold(deployment.label)} package: ${deployment.packageXmlFile} ...`));
+        const deployCommand = `sfdx force:source:deploy -x ${deployment.packageXmlFile}` +
+          ' --wait 60' +
+          ' --ignorewarnings' + // So it does not fail in for objectTranslations stuff
+          ` --testlevel ${testlevel}` +
+          (check ? ' --checkonly' : '') +
+          (debugMode ? ' --verbose' : '');
+          let deployRes ;
+        try {
+          deployRes = await execCommand(deployCommand, this, { output: true, debug: debugMode, fail: true });
+        } catch (e) {
+          const {tips} = analyzeDeployErrorLogs(e.stdout + e.stderr);
+          uxLog(this,c.red("Sadly there has been Deployment error(s)"));
+          uxLog(this,c.yellow(tips.map((tip:any) => c.bold(tip.label)+'\n'+tip.tip).join("\n\n")));
+          uxLog(this,c.yellow(c.bold(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the deployment issues :)`)));
+          throw new SfdxError('Deployment failure. Check messages above');
+        }
+        if (deployRes.status === 0) {
+          message = `[sfdx-hardis] Successfully ${check ? 'checked deployment of' : 'deployed'} sfdx project sources to Salesforce org`;
+          uxLog(this, c.green(message));
+        } else {
+          message = '[sfdx-hardis] Unable to deploy sfdx project sources to Salesforce org';
+          uxLog(this, c.red(deployRes.errorMessage));
+        }
+      }
+      // Deployment of type data import
+      if (deployment.dataPath) {
+        const dataPath = path.resolve(deployment.dataPath);
+        await importData(dataPath,commandThis,options);
+      }
+      // Wait after deployment item process if necessary
+      if (deployment.waitAfter) {
+        uxLog(this,`Waiting ${deployment.waitAfter} seconds after deployment according to deploymentPlan.json`);
+        await new Promise(resolve => setTimeout(resolve, deployment.waitAfter * 1000));
       }
       messages.push(message);
     }
@@ -76,25 +86,27 @@ async function buildDeploymentPackageXmls(packageXmlFile: string,check: boolean,
           packageXmlFile: mainPackageXmlCopyFileName,
           order: 0
         }
-        const packageXmlItems = [mainPackageXmlItem];
+        const deploymentItems = [mainPackageXmlItem];
         // Remove other package.xml items from main package.xml
-        for (const separatePackageXml of deploymentPlan.packages) {
-          separatePackageXml.packageXmlFile = path.resolve(path.join(path.dirname(deploymentPlanFile), separatePackageXml.packageXmlFile));
-          uxLog(this,c.cyan(`Removing ${separatePackageXml.packageXmlFile} content from main package.xml`));
-          const removePackageXmlCommand = 'sfdx essentials:packagexml:remove' +
-          ` --packagexml ${mainPackageXmlCopyFileName}` +
-          ` --removepackagexml ${separatePackageXml.packageXmlFile}` +
-          ` --outputfile ${mainPackageXmlCopyFileName}`;
-          await execCommand(removePackageXmlCommand, this, { fail: true, debug: debugMode });
-          packageXmlItems.push(separatePackageXml);
+        for (const deploymentItem of deploymentPlan.packages) {
+          if (deploymentItem.packageXmlFile) {
+            deploymentItem.packageXmlFile = path.resolve(path.join(path.dirname(deploymentPlanFile), deploymentItem.packageXmlFile));
+            uxLog(this,c.cyan(`Removing ${deploymentItem.packageXmlFile} content from main package.xml`));
+            const removePackageXmlCommand = 'sfdx essentials:packagexml:remove' +
+            ` --packagexml ${mainPackageXmlCopyFileName}` +
+            ` --removepackagexml ${deploymentItem.packageXmlFile}` +
+            ` --outputfile ${mainPackageXmlCopyFileName}`;
+            await execCommand(removePackageXmlCommand, this, { fail: true, debug: debugMode });
+          }
+          deploymentItems.push(deploymentItem);
         }
 
         // Sort in requested order
-        const packageXmlItemsSorted = sortArray(packageXmlItems, {
+        const deploymentItemsSorted = sortArray(deploymentItems, {
           by: ['order','label'],
           order: ['asc','asc']
         });
-        return packageXmlItemsSorted ;
+        return deploymentItemsSorted ;
     }
     // No transformation: return initial package.xml file
     return [
@@ -104,3 +116,4 @@ async function buildDeploymentPackageXmls(packageXmlFile: string,check: boolean,
         }
     ]
 }
+