+++ src/hooks/init/check-local-sfdx-hardis-files.ts
@@ -179,6 +179,7 @@ async function getHardisForceIgnoreContent() {
         
         "**/__tests__/**",
         "**/pubsub/**",
+        "**data/**/target/**",
         "**SfdxHardisDeferSharingRecalc**"
     ];
     return forceIgnoreContent;